Proper Implementation of Clamping
The previous implementation was flawed in that it did not look at vector magnitude, only directional component magnitude. this does not clamp.

Clamping checks the length of the intended vector against the max length vector. If the intended vector is longer than the max length, you divide the intended vector by its own magnitude, then multiply it by the magnitude of the max length allowed.